Split Symbol and Program to separate files in Rust
