+++ icicle/src/program_c_api.cpp
@@ -16,51 +16,6 @@ template <typename S> Program<S>* create_empty_program() {
 }
 
 extern "C" {
-  // Symbol functions
-  // Constructors
-  SymbolHandle CONCAT_EXPAND(FIELD, program_create_empty_symbol)() { return new Symbol<scalar_t>(); } // TODO rename symbol functions without program and split to a separate file
-  SymbolHandle CONCAT_EXPAND(FIELD, program_create_scalar_symbol)(const scalar_t* constant) {
-    return new Symbol<scalar_t>(*constant);
-  }
-  SymbolHandle CONCAT_EXPAND(FIELD, program_copy_symbol)(const SymbolHandle other) { 
-    return new Symbol<scalar_t>(*other);
-  }
-
-  // Destructor
-  eIcicleError program_delete_symbol(SymbolHandle symbol)
-  {
-    if (!symbol) { return eIcicleError::INVALID_POINTER; }
-    delete symbol;
-    return eIcicleError::SUCCESS;
-  }
-
-  void CONCAT_EXPAND(FIELD, program_set_symbol_as_input)(SymbolHandle this_symbol, int in_index) {
-    this_symbol->set_as_input(in_index);
-  }
-
-  SymbolHandle CONCAT_EXPAND(FIELD, program_inverse_symbol)(const SymbolHandle input) {
-    return new Symbol<scalar_t>(input->inverse());
-  }
-
-  SymbolHandle CONCAT_EXPAND(FIELD, program_assign_symbol)(SymbolHandle this_symbol, const SymbolHandle other) {
-    this_symbol->assign(other);
-    return this_symbol;
-  }
-
-  SymbolHandle CONCAT_EXPAND(FIELD, program_add_symbols)(const SymbolHandle op_a, const SymbolHandle op_b) {
-    return new Symbol<scalar_t>(op_a->add(*op_b));
-  }
-
-  SymbolHandle CONCAT_EXPAND(FIELD, program_multiply_symbols)(const SymbolHandle op_a, const SymbolHandle op_b) {
-    return new SymbolHandle(op_a->multiply(*op_b));
-  }
-
-  SymbolHandle CONCAT_EXPAND(FIELD, program_sub_symbols)(const SymbolHandle op_a, const SymbolHandle op_b) {
-    return new Symbol<scalar_t>(op_a->sub(*op_b));
-  }
-
-  // TODO separate files upto here
-
   // Program functions
   // Constructor
   ProgramHandle CONCAT_EXPAND(FIELD, program_create_empty_program)() { return create_empty_program<scalar_t>(); }
@@ -93,8 +48,4 @@ extern "C" {
     }
     program->generate_program(parameters_vec);
   }
-
-  int program_get_nof_vars(const ProgramHandle prog) { return prog->get_nof_vars(); } // TODO remove
-
-  void program_print_program(const ProgramHandle program) { program->print_program(); } // TODO remove, it's for internal debug
 }
