+++ wrappers/rust/icicle-core/src/lib.rs
@@ -10,6 +10,7 @@ pub mod poseidon;
 pub mod poseidon2;
 pub mod sumcheck;
 pub mod vec_ops;
+pub mod symbol;
 pub mod program;
 
 #[doc(hidden)]
